Revert "changed int x and y for scanner position, to double"
This reverts commit 0befc01284854f31228745436fd5c6889ad3c219.

+++ PARSE/Tracking/MeasurementLoader.xaml.cs
@@ -114,7 +114,7 @@ namespace PARSE
             }
         }
 
-        internal void capture(double x, double y, double angleXY, double angleZ)
+        internal void capture(int x, int y, double angleXY, double angleZ)
         {
             System.Diagnostics.Debug.WriteLine("Scan captured! END");
             tracker.Stop();
